refactored authentication routes and added roles and roleChecking for user (major)
