+++ components/user/userRouter.js
@@ -5,11 +5,4 @@ const { authCheck } = require('../../middlewares/authCheck');
 
 const router = express.Router();
 
-router
-  .post('/register', catchAsyncError(UserController.register))
-  .post('/signin', catchAsyncError(UserController.signIn))
-  .post('/reset-token', catchAsyncError(UserController.generateResetToken))
-  .post('/reset-password/:resetToken', catchAsyncError(UserController.resetPassword))
-  .post('/change-password', authCheck, catchAsyncError(UserController.changePassword));
-
 module.exports = router;
