+++ models/index.js
@@ -1,7 +1,8 @@
-const { User } = require('./user');
+const { User, roleTypes } = require('./user');
 const { School } = require('./school');
 
 module.exports = {
   User,
   School,
+  roleTypes,
 };
